fix(neural-networks): widen element-count products before they overflow int32
CodeQL cs/loss-of-precision:
- EmbeddingLayer.cs:856 (1161): vocab*dim*2 overflows for real LLM vocabularies
  (152064 x 8192 x 2 = 2.5e9), making the regularisation loss negative.
- AudioVisualEventLocalizationNetwork.cs:1322 (9504): frames x frame length overflows after
  ~346 1080p frames.
- SSMQuantizationHelper.cs:161 (15809): paramCount*bits overflows above 268M parameters at 8 bits,
  so EstimateMemorySavings reported negative bytes.

+++ src/NeuralNetworks/AudioVisualEventLocalizationNetwork.cs
@@ -1319,7 +1319,7 @@ public partial class AudioVisualEventLocalizationNetwork<T> : MultimodalModelLay
             }
         }
 
-        return _numOps.Divide(totalDiff, _numOps.FromDouble(frames.Count * frames[0].ToVector().Length));
+        return _numOps.Divide(totalDiff, _numOps.FromDouble((double)frames.Count * frames[0].ToVector().Length));
     }
 
     private Vector<T> EncodeTextDescription(string description)
